refactor(home): extract the fiat-totals warm-up into a hook
Pure extraction of the effect added in d74ed318 — HomeScreen now just
calls useWarmUpAccountsFiatTotals(), which reads the stores it needs
itself. Behavior unchanged, still covered by the HomeScreen warm-up
test.
